requirements.txt
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
asyncio-throttle>=1.0.0
aiohttp>=3.8.0
dataclasses-json>=0.6.0
typing-extensions>=4.5.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-json-logger>=2.0.0
matplotlib>=3.6.0
networkx>=3.0.0
pandas>=2.0.0
websockets>=11.0.0
scipy>=1.10.0